Raise combine/separate to Canvas
